Lots of investigations... and no conclusions
The code was modified to re-integrate the propagated (Souza’s) solution
instead of merely interpolating the solution into the linearly spaced
time array. In result, the value of P corresponding to the initial
solution has decreased substantially, but not enough. Big peaks of
error (P integrand) are visible when the thrust turns off and on again,
but no mistake was found (actually, a pretty serious mistake was found
in calcPhi, line 197/284, but it made no difference whatsoever.

Since the solution only becomes worse with the restoration cycles, it
seems the restoration algorithm does not deal well with abrupt changes
on the solutions… or there’s still a mistake somewhere in the code.

Anyway, I am too tired, and can’t look at the code anymore for today.