Restoration: major modifications, minor advances
Major modifications to restoration scheme, introduction of adapted
version of restoration (ODErest - for ODE deviations only) to try to
lower the values of components of P (integral and final condition)
separately. Results are not so great, though.

There was a mistake on the formulation: according to Miele, the
terminal of P that takes into account psi is actually the norm squared.