Lots of changes, no big results...
Still trying to figure out the grad problem. Actually, the grad seems
to be working ok, but the restorations after the first gradient step
are unable to bring the P value back to 1e-7.  Initially, it was
thought that the “step overdrive” (alfa>1) was leading the solution to
an “unrecoverable point”, but the solution behavior is very similar
without any “step overdrive”.

A major change was to restrict (in a hard-coded fashion, I must
confess) the pitch program to half a degree, and for more time, and
hence perform the pitch more smoothly. This does not impact the process
overall, so it was kept that way.

Another issue studied was control saturation. For that, a saturated
control problem was introduced in prob_teste. Apparently, both the
gradient and restoration are able to recover from saturated controls at
a wrong place, but a more “aggressively wrong” initial guess is still
to be tried.