Problem clarifications
Some tests performed with prob_test helped understand better the
problems faced by the program. In a problem whose solution involves
control saturation, an already saturated initial condition prevents the
convergence to the actual solution. Therefore, the initial guess
probably may not be OK for the algorithm to work on because the thrust
is most of the time saturated at both sides.

Another problem was discovered. There is a numerical problem with the
rocket equation that allows basculation of the rocket prior to any
attack angles. This numerical error is, unfortunately, essential to the
proper convergence of the initial guess in itsme.py…