Introducing sigmoid
In order to properly study the inefficiency of the grad-rest cycles,
the simple problem of optimal control was again used. The control
saturation was re-implemented with a sigmoid (hyperbolic tangent)
instead of a sine; the sine periodicity and zero derivative at +- .5*pi
are complications to the problem.

Some runs were performed, with and without “overdrive” in rest and grad
steps. It was discovered that the best performance happens with no
overdrive from grad.

The restoration still takes too many iterations to restore the P value
back to acceptable values. The corrections start to reveal high
frequency oscillations, which are inconsistent with the problem
dynamics. Perhaps these oscillations come from the numerical errors in
the A differential equation that are amplified because of the high K
values.

Hence, a new study will be performed to alter the “mu” initial
conditions, keeping the K values at “normal” levels and preventing high
frequency (numerical) disturbances from appearing.