Crucial changes to grad!
According to last commit, the grad algorithm was changed specifically
in the region of lambda initialization (mu). This modification was
tested in the test problem of the last commits, and the results are
amazing. Since no high frequency perturbations are introduced in grad
steps, only one restoration step is necessary to bring back the P
values to acceptable levels. Overdrive was allowed in calcStepGrad.

Now, this change will be applied to the main problem.