prob_rocket: sigmoid in control
In spite of the very promising results of the previous commit, applying
the hyperbolic tangent (sigmoid) to the controls in prob_rocket did not
yield good results. The algorithm takes too many restoration cycles to
correct the initial value of P provided by the initial guess (itsme).
This still has to be improved and investigated.

A different test problem (inverted pendulum) was conceived to test the
algorithm with a non-linear, unstable system. The algorithm worked
nicely with this problem. The first steps of grad kept step values low
(probably because of the initial parameter value), but the algorithm
converged very quickly.