sgra: high frequencies are hard to deal with
A test was performed with the “toy problem” (simple pendulum),
initializing the problem with an initial guess that contains high
frequency content. Indeed, the sgra algorithm does not manage to
converge with this high frequency content.

A somewhat desperate attempt was tried in the initMode=‘crazy’, with no
attack angle, and a constant moderate thrust. It did not work, quite
naturally.