Exponential damping and lots of debugging
It was discovered that the algorithm was excessively trying to make sharp turns in the begging of the flight, probably due to numerical errors. So, the dynamics was altered so that the gamma equation becomes more (artifically) rigid. The method is still not converging, though.

There are 3 serious problems:

1- High frequency noise is being inserted due to grad.  Apparently the solutions are not LI enough (ill-conditioned linear system?). Maybe initializing the lambda equations with different mu's... this still has to be looked into.

2- The grad still raises the P values, specially in the neighborhood of 0.912. The residue is due to the mass equation, which is strange. Is the system this non-linear?

3- The grad steps get "small" because of the initial error of the lambda, raising the Qx parcel of Q. Changing the N value switches the point of maximum Qx from the beggining of solution to middle (0.7).

In time: increasing N switches the point of maximum Qx but not the point of maximum P.

The code will be refactored to allow better debugging.