Some debugging upgrades
Some functionalities were added for debugging, such as optional plotting for grad, and an object for managing the overall run of the program. The overall method seems to be working, execution ended (with low thresholds). With more adequate thresholds, the solution does not converge at all.

Main refactoring of the program is still going to take place, with full objectification of the solution, including disk read and write actions for loading partial solutions.