Re-test Souza's new PR, calcStepGrad is no good anymore
Re-testing Souza's new PR: it turns out to yield trajectories
very close to the original (monostage) version of itsme.
Unfortunately, the newest changes to calcStepGrad seem to have
made it unsuitable for the rocket problem. Even with the cart
problem, the grad steps are too great, so more rest iterations
are necessary. This has to be thought over and dicussed with the
team for the msgra.