Rest is working... sort of
Latest changes to master branch were brought to this branch, in order
to avoid (or minimize) later problems.

A new toy model was introduced, namely, the probCart. It was actually
adapted for msgra by letting the initial conditions free, so it is a
multi-subarc instance with only 1 arc. The algorithm should apply,
anyway.

Some serious mistakes in the LMPBVP implementation were fixed, and
the restoration works... sort of. The P values decrease up to
˜9.0E-4, and stay there. This obviously still has to be investigated.