Manually updating Souza's last PR to master
Also, the calcStepGrad algorithm was changed to search a step by
minimizing I instead of Q. It seemed to work fine on probRock,
but it did not work on probCart because I became negative!
Besides, the code became a huge mess.

Therefore, a major overhaul will be performed in calcStepGrad,
probably involving creation of new objects.