calcStepGrad refactored, problems continue...
The calcStepGrad method was refactored, with the creation of new
object 'stepMan' (stepManager). It is possible now to perform the
step search with any objective funcion involving P, Q and I.

The program shows an old problem again, lots of restorations at a
certain point during the optimization. Therefore, some changes
are still needed in calcStepRest. This will be done in further
commits.