So far so good...
The algorithm worked fine with the last used value of k in objective
function, optimizing the solution up to 25.2% more payload, at a final
level of Q = 65.2 (Qu = 65.18). The program stopped at 1000 gradient
runs.

The program will now be changed to allow more gradient iterations,
and issue #31 will be addressed.