Fixed issue #33, and #31 partially
See the text of issue #31 (Saving plots to disk) for more details.

Also, fixed a minor (?) screw-up on calcStepRest.

Could not help but test the performance a little bit with probCart.
Unfortunately, this step-searching algorithm that minimizes
	Obj = I + k*P
is very "unrobust". If k is too low, then the steps get too high and
lots of restorations are needed to bring back the P levels. If k is
too high, the steps get too low and the solution doesn't go anywhere.
This is mainly because the algorithm is set to minimize Obj with 20%
alterations in the step.

It could be interesting to discuss a better search mechanism with
the team.