Grad improvements + convergence study
Many 'little' improvements were performed in grad_sgra, in the context
of the probCart problem. First of all, plotting the gradStep (why only
now?) added a lot of information. In the cart problem, no value of k
could be found to adequately allow convergence. Inspecting the gradient
step showed the step was too low at every iteration, for the P values
increased too quickly. This could be solved by increasing N; the
refinement was probably not good enough.

After that, many k values were tried. The table below shows the total
execution time for solving the cart problem completely in each case.

k    T[s]
1e-4 892.8
1e-5 445.4
1e-6 223.4
1e-7 122.6
1e-8 NoConv

As expected, the low values of k make the step size grow, for the step
search algorithm cares little for the increase in P. Trying k=1e-8
caused the solution to reach a P value so high that the restoration
was not able to completely restore, so it only kept recalculating
everytime.

The results show that the k value should be kept at a minimum, but not
so low as to allow this effect mentioned in the paragraph above.