Fixed issue #35
Now, with a simple try...except block, the plot saving commands do
not stop the flow of the main program, even if they fail.

Also, since for some reason the parallelism is not working in some
computers (Windows maybe?), there is a variable now to switch the
parallelization on or off.

On with the development!