Previous "issues" solved, new ones identified
As the team waits for the new itsme functionalities, the issues
were addressed, and solved. The program seems capable of
optimizing multistage rockets as long as the stages are all equal
in terms of Isp, structural efficiency and Thrust.

Another issue that reveals itself concerns a better organization of
the optimization outputs. The results are all "probRock_[something]",
and it would be better to refer them to the case (.its file) instead.

This is easier said than done because of the structure of the modules,
but it should be feasible anyway. It will be addressed next.