MAJOR improvements to program performance
Several improvements were made to increase performance of the program.
These are the main ones:

	- The "free" axis on calcGrads is now the time, this makes
	  calculations much more efficient
	- In calcStepGrad: no calculations of Q (unnecessary in
	  current objective design), and the search for step begins
	  with the previously used step, greatly reducing the number
	  of objective function calculation.
	- In calcStepRest: a simple check for P1<tol['P'] makes most
	  of the rest runs complete with a single step check.
	- In sgra: minor changes in LMPBVP functions save some
	  computational effort.

Running the program under these conditions changes the trajectory to
a less "upwards" and more "oblique" one, greatly reducing the losses
with respect to ideal Delta V (Tsiolkovsky).

Quite interestingly, the program was run with a rather unelegant
bypass, allowing greater thrusts and greater attack angles AFTER the
initial guess delivered by itsme. Under these conditions, the program
increases the payload (of a 2stage LV given by default2st.its) by
50%! This is quite impressive, and actually seems "too good to be
true". Hence, more tests must be concieved and run to check the
validity of these results.

There is a problem, however. The program does not converge (that is,
Q<tolQ), it only lowers the value of I, and also, somewhat unsteadily,
of Q, but it then reaches a point where more and more restorations are
needed to satisfy the P<tolP condition. This must be verified!