ProbCart investigations, and debug probRock
Some investigations were performed in probCart, altering the tolerances
for P and Q, and checking the relative difficulty of meeting these
requirements with different values of N. As expected, a lower value of N
makes a low tolP hard to reach. A high value of tolQ rather surprisingly
still yields good results, at least by comparing with the acutal optimal
solution. Low values of tolQ give out the same result observed in probRock,
that is, the solution reaches a point in which many restorations are made
but the tolP condition is not met.

The other "issue" investigated is the apparent insensitivity of the solution
on the coasting region, as if the algorithm never calculates a non-zero
correction for the thrust in the points in which it is already zero. This
makes sense from the point of view of the algorithm, because all the
corrections in control are minimized, so the algorithm would never "waste"
control effort in regions in which the control derivative phi_u is zero
(hence, the control effort is useless for minimizing error).

This has to be addressed, for the optimal solution (at least according to
Miele) involves a long coasting phase (zero thrust). A possibility is to
divide the flight in more subarcs than stages, allowing for an entire
subarc of coasting, as Souza suggested. This would also help for keeping
the N values low, by "refining" the time grid only in the more "sensitive"
regions.

Also, other alterations were made:

- Corrected mistakes in calcQ (Miele's revisor sucks)
- included a way to plot residuals of Qx, Qu, Qp, etc.
  (currently commented out: still has to be integrated with "debugMode")
- included a plot + save for lambda and the gradCorrection