More saturation tests and ideas
In order to properly test the effects of the saturating controls, two tests
were performed.

Firstly, probCart was initialized with a fully saturated forward control.
The result was that during the first restoration rounds, the time (pi)
became negative for the second subarc, and no recovery from this was
possible.

After that, probRock's initialization by itsme was "adulterated" by changing
the controls for no thrust level, from -5.0 to -2.5. This corresponds to
changing beta from .00004 to .007, or thrust from .04N to 7N (which is not
really "thrust free"...). The initial restorations were able to correct the
inconsistency produced by this "adulteration". The control also became less
unsensitive, as expected, but after 70 gradient iterations, the "intensive
restoration" cycles came back.