The longest run ever
An experiment was conducted, increasing N to 50000. A record setting
run of ~56h was performed! The same effect was observed: many res-
torations up to some point, resulting in near depletion of HD
because of the many .pkl files saved to disk. Nevertheless, the
solution reached a whopping 68.09% increase in payload capability,
over 300 gradient Iterations, with Q as:
Q = 9.1015E+03: Qx = 8.9655E+03, Qu = 1.3599E+02, Qp = 2.0044309E-02,
Qt = 7.7069E-26
	that is, Qx >> Qu!

To do next: remove the “many iterations routine” from interf, plot
and save integrand of Pint. Something very strange happens near the
start of the trajectory, for Qu and specially Qx present a spike. It
is probably due to lambda v. This should be investigated.
Later, maybe, change the calcGradStep to make it more like the search
suggested in Miele (2003).

The changes:
- changed the detailed plot of Q residuals for saving pdfs;
- added a generic plotSol method for sgra. (I had nothing to do!)

Also, a tragic error was found in LMPBVP. It will be corrected now.