More of the same: experiments and mistakes
Main experiment: new run, with the stupid mistake in sumIntFpi fixed.
Better performance, but the same final result; lots of restorations
A more careful analysis on the residuals showed a spike on Qx near the
very beginning of the trajectory, which may be due to the integration
scheme used. Down the rabbit hole we go...

I am seriously considering including many subarcs and maybe even
cheating on the control derivatives so that  the algorithm is "not
inclined" to changing the control profiles there. That may be special
useful for the vertical ascent phase.

changes:
- added external command for plotting (and saving) the details for Qx,
 Qu, etc
- added similar command for plotting and saving the Pint integrand.
- enabled parallelism to rest (I know, I'm an idiot. At  least now I
included a warning for not running grad/rest in non-parallel mode.)