Backwards Euler implicit: nice try
Interesting experiments were performed with integration methods for
propagating the EDOs. The main problem is still the cumulation of
Qx (specially in the region just after the "initial instants". The
solution ends up with more and more rest steps. Only a numerical
error could explain that symptom.

In other experiments, I tried to cheat in the initial and final
instants, so that the error computed by Pint ou Qx would be low by
design, but the error was just propagated forward or backward,
accordingly. Finally, the integration algorithm was changed to
"Backwards Euler Implicit", for it is a more stable method, but with
no sucess in long terms. Even with a relatively high value of k in
calcGradStep, the same overall behavior was detected, with a slow
decrease in Qu concurrent with a slow increase in Qx, with a great
rise in rest steps by grad step (called 'GRrate').

Other changes:
- Included arrays for calculating the aforementioned 'GRrate'
- Included general method for plotting Q residuals
- Updated probCart's plotSol method, for all modes (sol, corr, lambda).