More subarcs = less trouble? Probably!
Another subarc was inserted in the beginning of the trajectory, so
that the time step can be kept very low in the beginning of the
trajectory without increasing N enormously. This seems to be indeed
very promising for the Qx was not increased significantly near the
end as before. Maybe it would be interesting to add even more
subarcs in the beginning and in the end of the flight. For that,
though, it would be better to remove the "hardcodedness" of this
first subarc condition, to allow more flexibility and less code
alterations.

Main modifications: besides the modifications for the subarc
insertion mentioned above, the grad-rest rate plotter was included.
It can be seen that the GRrate explodes close to the "end" of the
simulation, causing the solution not to go anywhere. Perhaps a
control law involving the k value in grad and the GRrate might be
promising...