Accum. plot for P_int gives new insights
Back to the main problem of adjusting a suitable N for optimization,
plotting the accumulated integral for P_int gives more insight for
the points that contribute most for the final value of P_int.

(By the way, this technique should be used for other integrals, such
as Qx, Qu, etc. Maybe even refactor the function.)

Performing the analysis above, it can be seen that the final burn for
satelization also deserves better discretization, for the P_int gets
significantly high there.