Tackling first issues
I began to address the issues listed in the last commit. The overall
performance of the program is poor, since the grad step sizes are
too small. This is probably due to overly aggressive corrections
proposed by grad, as the trajectory given by itsme hits an
acceleration peak of ~30g's, when the limit used is 3g.

A modification in the penalty function was discussed with Muniz,
switching from a quadratic function (as proposed by Miele) to a
saturating curve such as a... hyperbolic tangent. Perhaps it will
have to be done at some point.

Main changes:

- A method for calculating the acceleration was provided;
- The methods for calculating f and its gradients were refactored
to yield less computational cost;
- A simple heuristic for choosing the penalty function constant Kpf
was discussed and implemented with Muniz.