P_int error visualization and other PF tests
Finally, a module for visualizing the error components in P_int,
namely, the vector (dx/dt-phi), was implemented. It shows that the
component which brings the greates ammount of error is, rather
surprisingly, the mass state! This was unexpected because the mass
equation is linear on beta, and independent of all other states.

Other tests on the penalty function were performed, and none have
showed good results. Maybe the initial guess is just too wrong in
terms of acceleration...