Mistakes and more experiments
Still trying to test the actual limitations of the acceleration
restriction, the program was run with a high value of acc_max,
so that the initial guess does not violate any conditions.

It still did not work, as the solution ended up hitting the
acc_max and then only restorations. Even after lots of errors
being discovered in calcGrads, it did not work.