Sliding down the brachistochrone!
In order to assess the capabilities of the SGRA program as is, a
new problem was included, namely, the classical brachistochrone
problem. The current implementation solves the problem in about 50
gradient iterations (almost no restoration is necessary), with
k=1e-5.

First tests show a great sensitivity on this constant k for grad
steps. This situation calls for more tests!

Changes are basically the obvious inclusion of the probBrac module,
as well as k alteration in grad_sgra and a small alteration (in
fact, adequation) of probCart for the new three-fold output of
calcI and calcF (for penalty function inclusion).