Interesting experiments yield good insights
Further experimenting with the brachistochrone problem showed
extreme sensitivity in the k values used for calcStepGrad. Even
for k values that resulted in convergence, there was a strange
oscillating pattern with the grad step. An investigation showed
that this oscillation was due to the k definition, as

k = 1e-5 * I0/P0.

When the calcStepGrad function was called with a very low value of
P0 (much lower than tol['P']), the k value was too large, and hence
the step is lower than it should. Because of this, the rest phase
was skipped, and the P value rose. In the next grad phase, P0 was
higher, yieldind a lower k and hence a higher step, and so on,
until P > tol['P'], when the restoration kicks in.

This effect was supressed changing the aforementioned equation from
P0 to tol['P'], but the multiplicative constant had to be changed,
of course. This led to a convergence study from

k = I0/tol['P'] to k = 1e-4 * I0/tol['P'].

The first one converges in a VERY slow and stable fashion, while the
second one could not be shown to converge, as it entered in a stange
limit cycle: due to the non-linearity of the equations, the calcula-
ted corrections resulted in an overshoot in the controls (compared
with the known optimal control profile, of course), which was over
correct with an undershoot, and so on. This unstabling process was
somewhat stabilized by the restoration, resulting in a cycle of
oscillations building up and than brought down by restoration. At
lower values of k, this oscillation led to a "permanent displace-
ment", a constant deviation from the known optimal solution that
the restoration could no longer smooth out.

This reminded me immediately of the weird pitch profile near the
beginning of the flight in probRock. At this point, I have no
choice other than switching back to probRock to do some tests!

Main changes:
- Substitution of P0 for tol['P'] in k definition in calcStepGrad
- Inclusion of positive tilting angles for the brachistochrone