Wrapping up before the hurricane
Continuing the work on the previous commit, experiments were performed
in order to test the insights. Another approach was tested, namely,
changing the objective function in calcStepGrad to

Obj = I + k * (P - tolP) * (P > tolP)

so that the P values do not offer any penalty on the grad step as long
as they are below the tolerance. Although that seems like a good idea,
it resulted on the same oscillating pattern reported above, as in some
grad iterations the penalty was active and in others, it was not.

Anyway, this will have to wait, because in the middle of the tests,
Muniz reported that THE FLOW OF THE MSGRA IS WRONG. That is, it does not
follow from Miele's original paper, or the one from 2003.

Hence, we decided to wrap things up here, synchronize this branch with
master, and start a new branch from there, since the changes will be
DRAMATIC.