Development of a more robust itsme
The basic idea is run itsme in two levels, one with a less accurate but robust model and with the results of this level, run the acurate but less robust model. The first level is represented by itsmeM.py and itsModelM.py.
Just to be sure about the update, I updated all files.