Saving log to file (iss #47)
The program now creates a folder and saves the log to a .txt file in
each run, complying with part of issue #47.

This was an unexpectedly complicated thing to do, for all the quick
solutions either wrote everything to a file or showed everything on
screen, but never both. I ended up having to alter pretty much all
the files that contained source code!

The solution used another object, called logger, that performs the
file saving operations without having to open and close the file all
 the time. This object was an attribute of both the iterations
manager and the running solution, but there were problems copying
objects with this logger as well as parallelizing in the LMPBVP
section. These problems were overcome, either passing the logger as
an argument in some situations, or binding it to another attribute
of another object as needed, e.g., in the case of the stepMan object
in calcStepGrad.

An instance of probBrac was fully run, and there were no errors (at
least in MacOS). Please note that issue #47 is not completely
solved; this is just a minor milestone to allow committing and
switching computers.