Saving .pdf files to case folder (iss #47)
Another part of issue #47 was solved: saving the .pdf files to the
folder where the case log is run. This was much easier.

For now, only the input files copy operation (for safe keeping) is
still to be done in order to fully solve the aforementioned issue.
This will be done after rewriting the code so that all the settings
are properly in the external configuration file.