Saving configuration file to folder (iss #47)
After a lot of work, "all" parameters for running sgra that are
problem-specific are in the configuration file, .its. After this
was implemented, saving the .its file to the folder was really
straightforward. The itsme.py had to be modified and was not
thoroughly tested, but I am pretty confident that it should work
normally because only a little configuration code was added. This
completely solves issue #47!

Also, the ancient "cost scaling factor", that had been set to 1.0
for the last 1454 versions, and that completely lost its purpose
when the penalty function was introduced to the cost, was finally
removed from the code.

All tests were made with probRock: All other problems still have
to be readequated to the current formulation.