New tricks to old dogs
All the old problems (probBrac, probCart, prob101) were readequated
to the new formulation. It is striking that the program as it
currently is, is actually quite uneffective for solving most
problems. Only in a rather narrow range of k (officially named
gradient step search objective modulation constant) the program
does converge to the optimal solution.

Luckily the flow will be restructured!