Parameter reload after initial guess
Finally, this functionality was added. Now the user can reset the
parameters of a solution after the initial guess was generated;
during the "Are these parameters OK?" session (checkPars). This
required editing all the problem instances one more time, but it
was worth it.

Also, the main function was modified, so that the user can select
the problem he/she wants to run, without having to edit the code
by uncommenting the problem instance, and commenting the others.
If the user is running the program from the terminal, instead of

$ python3 main.py

he/she can type

$ python3 main.py rocket
for probRock, or
$ python3 main.py brac
for probBrac, or
$ python3 main.py cart
for probCart... you get the idea. If the user does not type
anything, the algorithm goes on with probRock. For running with
spyder, the command is

runfile('.../SOAR/main.py', wdir='.../SOAR', args = 'cart')

instead of simply

runfile('.../SOAR/main.py', wdir='.../SOAR)'

The user can even use $ python3 main.py rock default2st.its
for setting the external input file. Check the code!