Minor changes in gitignore and .its defaults
Since the new changes by Souza, some files started to get tracked
by git, which is not desirable. This commit tries to fix this,
and also tries to clean-up other minor version control stuff.

Also, the default case for probRock was renamed from 'teste.its'
to 'probRock.its', in order to maintain consistency.