Refactoring and tuning 101 and cart
The first main change was the refactoring of the method
"loadParsFromFile": out of the specific problem instances to the
superclass (sgra).

Also, the parameters in defaults directory were tuned for problems
101 and cart; now both of them get solved rather quickly. In the
case of 101, the k value (gradStepSrchCte) was too high.

Another good bug removal was achieved switching from numpy.arange()
to numpy.linspace(); now there is no weird sensitivity to the N
values during creation of the self.t array.

I still have to tune the parameters for probBrac, though.