Tuning probBrac and quickfix for Windows issue
Firstly, I confess that I screwed up in the Windows overriding
option 2 commits ago. Sorry. Now it should be working.

Also, the probBrac has been tuned. There are lots of issues to
discuss, because the algorithm becomes unstable if the value of
k (gradStepSrchCte) is too low; this should not happen anymore.
The team shall discuss this later.