Problems with the new program flow
There was a problem in the grad adaptation: P value was calculated
for the old (pre-correction) solution, instead of the new one.
Because of this, there was no restoration increasing the value of
I, and hence the "grad step retry" part was never active.

After solving this problem, the method calcStepGrad was shown to be
ill-written, for the same value of step was chosen over and over
again. This will be corrected now.