Simplest adaptation to calcStepGrad: fail!
First of all, there is also a stupid bug fix here: self.method()
instead of sol.method(). Very stupid indeed. Now for the serious
part.

The simplest adaptation possible to calcStepGrad was tried here.
Basically, there is a new flag to inform calcStepGrad if the
program already tried to do a gradient phase with that correction
or not. If it has already tried and failed, then a smaller step
size must be found. The program reduces alfa by 10% and tries
again, because the objective function lowering condition is also
satisfied.

Naturally, this approach did not work. If the gradStepSrchCte is
low enough to allow big steps, then the program begins refusing
the gradient step after restorations, as planned, but in the next
phase it tries again a bigger value of alfa (because this lowers
the objective function), which then gets refused many times...
this greatly decreases the program performance.

We must redesign the calcStepGrad method entirely. Probably the
step "overdrive" will have to be limited, because otherwise the
program will keep trying big steps, only to be refused after the
restoration.

One issue remains: the Original Miele Algorithm₢ does propose
searching the gradient step by minimizing J with a fixed P
restriction instead of minimizing I+k*P. The time has come for a
big decision: switch to Miele's recomendation or keep trying as
we are?