(Unsuccessful) experiments with calcStepGrad
Two simple approaches were tried in order to overcome the problems
stated in the previous commit.

In the first attempt, disabling the "step overdrive" completely
(that is, prohibiting the program from returning step values greater
than the initial value provided) created the expected effect:
steady but unfeasibly slow convergence, for sometimes the program
needs to lower the step value and it is unable to increse it
afterwards.

In the second attempt, the step raise was limited to 1.1 times the
previous value. This worked better than before at first, but then
the grad step rejections began. This is a problem...