Inclusion of probLQR to study MSGRA
The motivation behind this was to create a new problem, very
"tunable", for testing the sensitivity of the MSGRA program to
unstable eigenvalues, or initial conditions, etc, in a linear
problem.

The first sad truth was discovering that free final time problems
all become non-linear in SGRA. Another unfortunate fact is the
apparent sensitivity to initial guesses, in which the final
solution is different depending on the initial guess! Maybe that
is because of the low tolerance values of tolP and tolQ, but I
do not know for sure.