Studies in LQR yield good insights
First of all, I changed the problem for a fixed time version of
LQR. This should make the problem completely linear, and hence,
solvable in one (1) gradStep. It turns out there were numeric
problems preventing the convergence. This is odd because both
the eigenvalues were negative (-2.0 and -3.0). Checking the
solutions, there were strange high frequency oscillations on the
lambda and hence on the controls.

In order to test this hypothesis, the eigenvalues were changed to
more "intense" ones, more specifically, -5.0 and -10.0, and the
oscillations worsened. Based on these issues, the most likely
cause of the problem is numerical instability. This was confirmed
by increasing N (a lot, to 20000) and the problem became less
intense.

This is a big problem! Eigenvalues with high absolute values are
numerically dangerous, for either x equation or the lambda
equation, because the eigenvalues for lambda are the opposite of
the ones for x (that is, for the system).

I will try to insert another solver in sgra/LMPBVP; let's hope
this works.