Matrix exponentiation as solver: disaster
Yes. The matrix exponentiation was implemented as a new solver
option, trying to yield much less error, as the solution is almost
analytical (assuming the controls are constant during the
integration interval, naturally).

The result was a disaster. The computational cost incresed
monstrously, with no great reduction on the numeric error. Even
worse, it fails completely (i.e., stops the program execution) in
problems where there are null eigenvalues, e.g., probCart. It is,
at least in the current form, a complete disaster.

So much for all the hope...