Experimenting with solvers
Following Muniz's recommendation, I decided to study prob9_1. It
struck me that Miele did write a lot of specific information con-
cerning this proble in his paper, so we could try to match the
performance of the code. His claims are VERY unreliable--he says
that the SGRA manages to get P<1e-8 in four restorations with N=50!
Nevertheless I started matching all the settings to test his state-
ments.

First of all, the integrations were switched to Simpson's method,
increasing the computational cost just a little. Then, the solvers.
Miele claims to use "a modified Hamming predictor-corrector scheme,
with a special Runge-Kutta procedure to start the integration
routine", pointing to a book by RALSTON, A., "Numerical Integration
 methods for the Solution of Ordinary Differential Equations,
Mathematical Methods for Digital Computers, Vol. 1, Edited by A.
Ralston and H. S. Will, John Wiley and Sons, New York, 1960."
I could not find that reference, nor reproduce the Hamming scheme.
It is important to mention, however, that with Backwards Euler Im-
plicit, N=1001 is still not capable of meeting P<1e-8, in ANY
number of restorations...