Miele's results in prob9-1 were matched!
After reading many references concerning numerical methods for
ODEs, I decided to stay with the trapezoidal (implicit) method. It
is, according to a theorem by Dahlquist, the "best" A-stable method
(actually, LMM). Griffiths and Higham's "Numerical Methods for Or-
dinary Differential Equations" proved to be a great reference in-
deed. I had already tried that method before, so something else had
to be altered as well.

Thanks to a series of very interesting mandatory lectures today, I
deduced the expression of the error (phi-dx/dt) after running a res-
toration round. It turns out the error does get reduced a whole or-
der of magnitude when the method used is Euler for the first point
and "leapfrog" for the middle points.

I then concluded I needed to propose a different (but still consis-
tent) method for calculating the error; one compatible with the tra-
pezoidal scheme. I arrived at a fairly elegant method, defined by:

(e_k + e_k+1) = (phi_k+phi_k+1) - (x_k+1-x_k)/(2*dt) , 0<=k<=N-2

The previous relation yields N-1 equations; so I decided to divide
the error in the first equation (e_0+e_1) equally between e_0 and
e_1. The other error terms can be calculated with the equations.

It worked PERFECTLY! Even lowering the N value to 51, the P values
dropped to < 1.5e-16 after only 4 restorations! This is actually
better than Miele's results. To be fair, I still to this point have
no idea of the "Hamming's modified predictor-corrector method with
a special Runge-Kutta procedure to start the integration routine".
But, with even better results, I believe we are ready to proceed.

These are glorious times.