Changed lambda derivative calculations
There is a lambda (co-state) derivative calculation in one of the
error components of Q:
Qx = dlam - fx + phix.transpose().dot(lam)

This was changed in a similar fashion as the P residual calculation,
in order to become more compatible with the trapezoidal rule.

For all intents, this prob9-1 was run under this configuration and
converged in 8 gradient iterations. Things are looking good!