Properly recording the histories
The methods concerning the update and exhibition of the histories
(histP, histQ, histI, histGradStep, GRrate, etc)
have been (finally) updated for the new flow of the program.

When testing these methods, a weird situation occurred. In a grad
step, the solution failed to meet the I<I_old condition, leading
to a grad reject. The problem is that the Q<tolQ condition was met
in this step! Luckily in this case, the next try led to a grad
acceptance, but there could have been many rejections, and many
restorations without necessity, because the Q condition, which is
much more important, had already been met.

I will try to solve this now.